Minecraft finished upload

+++ week-03/day-05/Exercises/src/Minecraft/Gold.java
@@ -1,8 +1,9 @@
 package Minecraft;
 
-public class Gold extends Block{
-    public static String [] blocks = {"Glowstone", "Glass"};
-    public Gold(double lightTransmission) {
+public class Gold extends Block {
+    public static String[] blocks = {"Glowstone", "Glass"};
+
+    public Gold(int lightTransmission) {
         super("Gold", "rgb: 255,215,0", false, blocks);
         setLightTransmission(lightTransmission);
     }
